LunchTime_Python/lunchtime/requirements.txt
psycopg2-binary == 2.9.3
Django == 4.0.3
djangorestframework == 3.13.1
Pillow == 9.1.0
django-rest-passwordreset
django-environ
